rename async module to asynchronous for compatibility w/ python 3.7 and backwards compatibility with <3.7
