+++ agavepy/__init__.py
@@ -1,1 +1,6 @@
 from .agave import Agave
+
+import sys
+
+from agavepy import asynchronous
+sys.modules['agavepy.async'] = asynchronous
