+++ agavepy/tests/test_agave.py
@@ -27,7 +27,7 @@ import pytest
 import requests
 
 import agavepy.agave as a
-from agavepy.async import AgaveAsyncResponse
+from agavepy.asynchronous import AgaveAsyncResponse
 from . import testdata
 
 HERE = os.path.dirname(os.path.abspath(__file__))
